+++ engine/src/com/twinsoft/convertigo/engine/admin/services/studio/dbo/Remove.java
@@ -32,7 +32,7 @@ import com.twinsoft.convertigo.engine.admin.services.at.ServiceDefinition;
 import com.twinsoft.convertigo.engine.admin.services.studio.ngxbuilder.BuilderUtils;
 
 @ServiceDefinition(name = "Remove", roles = { Role.WEB_ADMIN,
-		Role.PROJECT_DBO_VIEW }, parameters = {}, returnValue = "")
+		Role.PROJECT_DBO_CONFIG }, parameters = {}, returnValue = "")
 public class Remove extends JSonService {
 
 	@Override