+++ engine/src/com/twinsoft/convertigo/engine/admin/services/studio/dbo/Save.java
@@ -31,7 +31,7 @@ import com.twinsoft.convertigo.engine.admin.services.ServiceException;
 import com.twinsoft.convertigo.engine.admin.services.at.ServiceDefinition;
 import com.twinsoft.convertigo.engine.sync.SharedWorkspaceSyncManager;
 
-@ServiceDefinition(name = "Save", roles = { Role.WEB_ADMIN, Role.PROJECT_DBO_VIEW }, parameters = {}, returnValue = "")
+@ServiceDefinition(name = "Save", roles = { Role.WEB_ADMIN, Role.PROJECT_DBO_CONFIG }, parameters = {}, returnValue = "")
 public class Save extends JSonService {
 
 	@Override